--- cla-bot/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "cla-bot-lambda",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "^2.149.0",
-    "handlebars": "^4.0.10",
-    "is_js": "^0.9.0",
-    "jsonwebtoken": "^7.4.0",
-    "parse-github-url": "^0.3.2",
-    "request": "^2.81.0",
-    "uuid": "^3.1.0"
-  }
-}
