+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "execute:bot": "serverless invoke local --function cla-bot --path cla-bot/dev/event.json --stage dev",
     "test": "npm run lint && jasmine",
-    "lint": "eslint --ignore-path .gitignore \"cla-bot/*.js\" \"spec/*.js\"",
-    "postinstall": "cd cla-bot && npm i"
+    "lint": "eslint --ignore-path .gitignore \"src/*.js\" \"spec/*.js\""
   },
   "author": "",
   "license": "ISC",
@@ -26,5 +25,14 @@
     "jasmine-spec-reporter": "^4.1.0",
     "mock-require": "^2.0.2",
     "serverless": "^1.27.3"
+  },
+  "dependencies": {
+    "aws-sdk": "^2.149.0",
+    "handlebars": "^4.0.10",
+    "is_js": "^0.9.0",
+    "jsonwebtoken": "^7.4.0",
+    "parse-github-url": "^0.3.2",
+    "request": "^2.81.0",
+    "uuid": "^3.1.0"
   }
 }
