+++ save-secrets.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-tar cvf secrets.tar clabot-dev-integration-key.pem clabot-integration-key.pem cla-bot/serverless.env.yml
+tar cvf secrets.tar clabot-dev-integration-key.pem clabot-integration-key.pem src/serverless.env.yml
 travis encrypt-file secrets.tar
