+++ src/contributionVerifier.js
