+++ src/requestAsPromise.js
